+++ packages/terra-popup/CHANGELOG.md
@@ -8,6 +8,7 @@ Unreleased
 
 ### Fixed
 * Lint errors regarding file path resolution
+* Popup Content will clone prop to indicate missing close button
 
 4.9.0 - (July 26, 2018)
 ------------------